refactor: extract server manager
